+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Supported compact mode for article list to reduce visual clutter. (#403)
 - Enhanced image gallery with support for multi-image articles and better navigation. (#457)
 - Supported Anthoropic and DeepSeek AI services.
+- Supported hiding text overlay in image gallery view. (#486)
 
 ### Changed